+++ configurator/TemplateConfigurator.rb
@@ -71,6 +71,7 @@ module Pod
       file_names.each do |file_name|
         text = File.read(file_name)
         text.gsub!("${POD_NAME}", @pod_name)
+        text.gsub!("PROJECT", @pod_name)
         text.gsub!("${REPO_NAME}", @pod_name.gsub('+', '-'))
         text.gsub!("${USER_NAME}", user_name)
         text.gsub!("${USER_EMAIL}", user_email)
@@ -80,7 +81,13 @@ module Pod
       end
     end
 
+
+
     def rename_template_files
+      # Move schemes
+      FileUtils.mv "PROJECT.xcodeproj/xcshareddata/xcschemes/PROJECT.xcscheme", "PROJECT.xcodeproj/xcshareddata/xcschemes/#{pod_name}.xcscheme"
+
+      # Move project files
       FileUtils.mv "PROJECT.podspec", "#{pod_name}.podspec"
       FileUtils.mv "PROJECT.xcodeproj", "#{pod_name}.xcodeproj"
       FileUtils.mv "PROJECT.xcworkspace", "#{pod_name}.xcworkspace"
