--- templates/objective-c/Example/PROJECT/CPDAppDelegate.h
@@ -1,15 +1,0 @@
-//
-//  CPDAppDelegate.h
-//  PROJECT
-//
-//  Created by PROJECT_OWNER on TODAYS_DATE.
-//  Copyright (c) TODAYS_YEAR PROJECT_OWNER. All rights reserved.
-//
-
-@import UIKit;
-
-@interface CPDAppDelegate : UIResponder <UIApplicationDelegate>
-
-@property (strong, nonatomic) UIWindow *window;
-
-@end
