--- templates/objective-c/Example/PROJECT/CPDViewController.h
@@ -1,13 +1,0 @@
-//
-//  CPDViewController.h
-//  PROJECT
-//
-//  Created by PROJECT_OWNER on TODAYS_DATE.
-//  Copyright (c) TODAYS_YEAR PROJECT_OWNER. All rights reserved.
-//
-
-@import UIKit;
-
-@interface CPDViewController : UIViewController
-
-@end
