--- templates/objective-c/Example/PROJECT/CPDViewController.m
@@ -1,29 +1,0 @@
-//
-//  CPDViewController.m
-//  PROJECT
-//
-//  Created by PROJECT_OWNER on TODAYS_DATE.
-//  Copyright (c) TODAYS_YEAR PROJECT_OWNER. All rights reserved.
-//
-
-#import "CPDViewController.h"
-
-@interface CPDViewController ()
-
-@end
-
-@implementation CPDViewController
-
-- (void)viewDidLoad
-{
-    [super viewDidLoad];
-	// Do any additional setup after loading the view, typically from a nib.
-}
-
-- (void)didReceiveMemoryWarning
-{
-    [super didReceiveMemoryWarning];
-    // Dispose of any resources that can be recreated.
-}
-
-@end
