--- templates/objective-c/Example/PROJECT/PROJECT-Prefix.pch
@@ -1,16 +1,0 @@
-//
-//  Prefix header
-//
-//  The contents of this file are implicitly included at the beginning of every source file.
-//
-
-#import <Availability.h>
-
-#ifndef __IPHONE_5_0
-#warning "This project uses features only available in iOS SDK 5.0 and later."
-#endif
-
-#ifdef __OBJC__
-    @import UIKit;
-    @import Foundation;
-#endif
