--- templates/objective-c/Example/PROJECT/en.lproj/InfoPlist.strings
@@ -1,1 +1,0 @@
-/* Localized versions of Info.plist keys */
