--- templates/objective-c/Example/PROJECT/main.m
@@ -1,17 +1,0 @@
-//
-//  main.m
-//  PROJECT
-//
-//  Created by PROJECT_OWNER on TODAYS_DATE.
-//  Copyright (c) TODAYS_YEAR PROJECT_OWNER. All rights reserved.
-//
-
-@import UIKit;
-#import "CPDAppDelegate.h"
-
-int main(int argc, char * argv[])
-{
-    @autoreleasepool {
-        return UIApplicationMain(argc, argv, nil, NSStringFromClass([CPDAppDelegate class]));
-    }
-}
