--- templates/objective-c/Example/Podfile
@@ -1,11 +1,0 @@
-source 'https://github.com/CocoaPods/Specs.git'
-use_frameworks!
-
-target '${POD_NAME}_Example' do
-  pod '${POD_NAME}', :path => '../'
-
-  target '${POD_NAME}_Tests' do
-    inherit! :search_paths
-    ${INCLUDED_PODS}
-  end
-end
