--- templates/objective-c/Example/Tests/Tests-Prefix.pch
@@ -1,7 +1,0 @@
-//  The contents of this file are implicitly included at the beginning of every test case source file.
-
-#ifdef __OBJC__
-
-  ${INCLUDED_PREFIXES}
-
-#endif
