--- templates/objective-c/Example/Tests/Tests.m
@@ -1,9 +1,0 @@
-//
-//  PROJECTTests.m
-//  PROJECTTests
-//
-//  Created by PROJECT_OWNER on TODAYS_DATE.
-//  Copyright (c) TODAYS_YEAR PROJECT_OWNER. All rights reserved.
-//
-
-${TEST_EXAMPLE}
