+++ templates/swift/PROJECT.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:${POD_NAME}.xcodeproj">
+      location = "group:PROJECT.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Example/iOS Example.xcodeproj">
