+++ templates/swift/Package.swift
@@ -1,5 +1,5 @@
 import PackageDescription
 
 let package = Package(
-  name: "FDChessboardView"
+  name: "PROJECT"
 )
